refactor(colorscheme): Modularize configuration of the editor colorscheme
Before, every colorschemes' options were placed in the
colorscheme.lua file. This made such every new colorscheme that was not
configured the same way as the existing ones would have to be set in a
special file, and changing the colorscheme required to configure
manually every detail.

This implementation lets me define the options of a colorscheme in a
file that is read like a module, setting the options and scheme
appropriately.